Move files
Moved IPAddressHelper and NetworkInfo classes, along with their tests, into new Converters namespaces and folders for better organization. Updated all relevant namespaces and using directives.
